# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Map Tracker App",
        "contact": {
            "name": "Ahmed Adel",
            "url": "https://www.linkedin.com/in/ahmed-mahmoud-168a8194/",
            "email": "adesouky@cat.com.eg"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "Ahmed"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "email@gmail.com"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "required": true,
                        "type": "file",
                        "format": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "ahmed@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "402": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout currently logged in user",
                "responses": {
                    "200": {
                        "description": "message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "402": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "refreshes expired token",
                "responses": {
                    "200": {
                        "description": "message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "402": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/check/{user}/add": {
            "post": {
                "tags": [
                    "Check in"
                ],
                "summary": "logout currently logged in user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer",
                        "default": "1"
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "Jeddah"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "Jeddah, Tihamah, Saudi Arabia\t"
                    },
                    {
                        "name": "latitude",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "21.543333"
                    },
                    {
                        "name": "longitude",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "39.172779"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "402": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}